+++ Laba_1/11.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 
 int main() {
-    double a, b, c;
+    float a, b, c;
 
     // Ввод коэффициентов a, b и c
     std::cout << "Введите коэффициент a: ";
@@ -14,16 +14,16 @@ int main() {
     std::cout << "Введите коэффициент c: ";
     std::cin >> c;
 
-    double discriminant = b * b - 4 * a * c;
+    float discriminant = b * b - 4 * a * c;
 
     if (discriminant > 0) {
-        double root1 = (-b + std::sqrt(discriminant)) / (2 * a);
-        double root2 = (-b - std::sqrt(discriminant)) / (2 * a);
+        float root1 = (-b + std::sqrt(discriminant)) / (2 * a);
+        float root2 = (-b - std::sqrt(discriminant)) / (2 * a);
 
         std::cout << "Два корня: " << root1 << " и " << root2 << std::endl;
     }
     else if (discriminant == 0) {
-        double root = -b / (2 * a);
+        float root = -b / (2 * a);
         std::cout << "Один корень: " << root << std::endl;
     }
     else {
