+++ Laba_1/20.cpp
@@ -7,7 +7,7 @@ int main() {
     std::cout << "Введите число N: ";
     std::cin >> N;
 
-    double log_base_22 = log(N) / log(22);
+    float log_base_22 = log(N) / log(22);
 
     if (std::floor(log_base_22) == log_base_22) {
         std::cout << N << " является степенью числа 22." << std::endl;
