+++ Laba_1/24.cpp
@@ -7,18 +7,18 @@ int main() {
     std::cout << "Введите координаты вектора (x y z): ";
     std::cin >> x >> y >> z;
 
-    double length = std::sqrt(x * x + y * y + z * z);
-    double scalarProduct = x + y + z;
-    double X = y - z;
-    double Y = z - x;
-    double Z = x - y;
-    double sumX = x + x;
-    double sumY = y + y;
-    double sumZ = z + z;
+    float length = std::sqrt(x * x + y * y + z * z);
+    float scalarProduct = x + y + z;
+    float X = y - z;
+    float Y = z - x;
+    float Z = x - y;
+    float sumX = x + x;
+    float sumY = y + y;
+    float sumZ = z + z;
 
-    double diffX = x - x;
-    double diffY = y - y;
-    double diffZ = z - z;
+    float diffX = x - x;
+    float diffY = y - y;
+    float diffZ = z - z;
 
     std::cout << "Длина вектора: " << length << std::endl;
     std::cout << "Скалярное произведение векторов: " << scalarProduct << std::endl;
