+++ Laba_1/6_7.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 
 int main() {
-    double a, b;
+    float a, b;
 
     std::cout << "Введите первое число: ";
     std::cin >> a;
@@ -10,8 +10,8 @@ int main() {
     std::cout << "Введите второе число: ";
     std::cin >> b;
 
-    double average_arithmetic = (a + b) / 2;
-    double average_geometric = std::sqrt(a * b);
+    float average_arithmetic = (a + b) / 2;
+    float average_geometric = std::sqrt(a * b);
 
     std::cout << "Среднее арифметическое: " << average_arithmetic << std::endl;
     std::cout << "Среднее геометрическое: " << average_geometric << std::endl;
